\name{logLik.fitmixst}
\alias{logLik.fitmixst}
\title{Loglikelihood for fitmixst}
\usage{
  ##S3 method for class 'fitmixst' logLik(object,...)
}
\arguments{
  \item{object}{of the fitmixst class.}
}
\value{
  the loglikhood of the object
}
\description{
  Loglikelihood for fitmixst
}
\keyword{likelihood}
\keyword{log}

