\name{print.fitmixst}
\alias{print.fitmixst}
\title{print for fitmixst}
\usage{
  ##S3 method for class 'fitmixst' print(object,...)
}
\arguments{
  \item{object}{of the fitmixst class.}
}
\value{
  the loglikhood of the object
}
\description{
  print for fitmixst
}
\keyword{object}
\keyword{printed}

