\name{summary.fitmixst}
\alias{summary.fitmixst}
\title{print for fitmixst}
\usage{
  ##S3 method for class 'fitmixst' summary(object,...)
}
\arguments{
  \item{object}{of the fitmixst class.}
}
\value{
  summary of the object
}
\description{
  print for fitmixst
}
\keyword{object}
\keyword{of}
\keyword{summary}
\keyword{the}

